Fix: minor changes like changing color
I fixed the color to match the one from the design and also I made some
changes to the projects section, changing the spacing for the icons and
the content of each projects. I am having trouble with the grid for the
project container. The main content is in progress and is mess. But I
will fix it.